+++ av1/common/common_data.h
@@ -79,7 +79,7 @@ static const uint8_t num_pels_log2_lookup[BLOCK_SIZES_ALL] = {
 
 #if CONFIG_CWP
 // Supported weighting factor for compound weighted prediction
-static const int cwp_weighting_factor[2][MAX_CWP_NUM] = {
+static const int8_t cwp_weighting_factor[2][MAX_CWP_NUM] = {
   { 8, 12, 4, 10, 6 },
   { 8, 12, 4, 20, -4 },
 };
